Update csvUpload.php
This is simple code to upload CSV data into Mysql Database, You can use this code as Core and Create a Plugin or Apps where one can upload the CSV in Plugin or Apps and it will upload as New Entries Or Create a condition that gives the User to Select Between New Entries or Update Old Ones. In Any Case, if you Feel like This Code Can Help You. Please Let Me Know.